MINOR: Add language support
Hi Thierry

In the past I've used the pla tool for some (french) projects.
The nice thing is that you can script (and thus version control) the project plan.

This time I am working on a project with english speaking stakeholders.
I needed a way to change the months.
I came up with this solution and I want to give it back to you.

I hope that it suits your standards.

Peter Reijnders